requirements.txt
# Interface graphique
PySide6>=6.5.0
qtpy>=2.3.0

# Imagerie médicale
pydicom>=2.3.0
# nibabel>=4.0.0
# SimpleITK>=2.1.0
# opencv-python>=4.7.0
Pillow>=9.0.0
# imageio>=2.25.0
# scikit-image>=0.19.0

# IA et Deep Learning (Désactivé pour économiser de l'espace - 4GB limit)
# torch>=2.0.0
# torchvision>=0.15.0
# monai>=1.2.0
# ultralytics>=8.0.0  # YOLOv8
# segmentation-models-pytorch>=0.3.0
# onnxruntime>=1.14.0
# onnx>=1.13.0

# Visualisation 3D
pyvista>=0.38.0
vtk>=9.2.0
# matplotlib>=3.5.0
# plotly>=5.13.0

# Analyse de données
numpy>=1.21.0
# pandas>=1.5.0
# scipy>=1.9.0
# scikit-learn>=1.2.0

# Rapports et export
# reportlab>=3.6.0
# Jinja2>=3.1.0
# python-docx>=0.8.0

# Utilitaires
# tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0